user delete for admin

+++ api/controllers/user.controller.js
@@ -47,10 +47,10 @@ export const UpdateUser = async (req, res, next) => {
 
 
 export const DeleteUser = async (req,res,next)=>{
-  if(req.user.id !== req.params.userId) return next(errorHandler(401, 'You can only delete your own account'))
+  if(!req.user.isAdmin && req.user.id !== req.params.userId) return next(errorHandler(401, 'You can only delete your own account'))
     try {
       await User.findByIdAndDelete(req.params.userId)
-      res.clearCookie('access_token');
+      if(req.user.id === req.params.userId) res.clearCookie('access_token');
       res.status(200).json({message : "user deleted"})
     } catch (error) {
       next(error)
